Updated NuGet package
Handle Pulsar Token bug in affected versions

The commit updates the used images of ToxiProxy and Pulsar in tests from versions 2.7.0 and 3.1.3 to 2.9.0 and 3.2.3 respectively to keep up with the latest versions.

The issue arises because Apache Pulsar incorrectly treats seconds as milliseconds in specific versions. To address this, we multiply the value by 1000 when using affected versions.

Finally, I added a missing cancellation token in ConnectionPool.